LayoutTests/fast/history/multiple-back-forward-navigations-expected.txt

Tests that pageshow / pagehide events are reliably fired when navigating back and forth several times.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* pageshow - persisted: false
PASS pagehideCount is 0
PASS event.persisted is false
* pagehide - persisted: true
* pageshow - persisted: true
PASS pagehideCount is 1
PASS event.persisted is true
* Navigating forward
* pagehide - persisted: true
* pageshow - persisted: true
PASS pagehideCount is 2
PASS event.persisted is true
* Navigating forward
* pagehide - persisted: true
* pageshow - persisted: true
PASS pagehideCount is 3
PASS event.persisted is true
PASS successfullyParsed is true

TEST COMPLETE
